Migrate opcache check to new SetupCheck API
