+++ core/js/setupchecks.js
@@ -240,18 +240,6 @@
 							type: OC.SetupChecks.MESSAGE_TYPE_ERROR
 						});
 					}
-					if(data.OpcacheSetupRecommendations.length > 0) {
-						var listOfOPcacheRecommendations = "";
-						data.OpcacheSetupRecommendations.forEach(function(element){
-							listOfOPcacheRecommendations += '<li>' + element + '</li>';
-						});
-						messages.push({
-							msg: t('core', 'The PHP OPcache module is not properly configured. See the {linkstart}documentation ↗{linkend} for more information.')
-								.replace('{linkstart}', '<a target="_blank" rel="noreferrer noopener" class="external" href="' + OC.theme.docPlaceholderUrl.replace('PLACEHOLDER', 'admin-php-opcache') + '">')
-								.replace('{linkend}', '</a>') + '<ul>' + listOfOPcacheRecommendations + '</ul>',
-							type: OC.SetupChecks.MESSAGE_TYPE_INFO
-						});
-					}
 					if(!data.isSettimelimitAvailable) {
 						messages.push({
 							msg: t('core', 'The PHP function "set_time_limit" is not available. This could result in scripts being halted mid-execution, breaking your installation. Enabling this function is strongly recommended.'),
